feat: make persistentReducedStream register streams
